Import the LGPL 2.1 license text
While there is no code in the repo yet, add the LGPL 2.1 license
text as a placeholder so that GitLab identifies a license for
the project. The license can be switched at a later date, if we
ever go forward with the project and decide LGPL 2.1 was not
the optimal choice.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>